feat: Kanban board para flujo de publicacion de contenido
Anadido tablero kanban para publicacion de nuevo contenido.

+++ modulos/Posts/views.py
@@ -1,15 +1,19 @@
-from django.contrib.auth.models import Group
-from django.contrib.auth.decorators import login_required
-from django.http.response import HttpResponseBadRequest
-from django.shortcuts import HttpResponse, get_object_or_404, redirect, render
-from modulos.Authorization.roles import ADMIN
-from modulos.Authorization.decorators import permissions_required
-from modulos.Authorization.permissions import (
-    POST_CREATE_PERMISSION,
-    POST_DELETE_PERMISSION,
-    POST_EDIT_PERMISSION,
-)
+from datetime import timedelta
 
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import Group
+from django.http.response import HttpResponseBadRequest, HttpResponseForbidden
+from django.shortcuts import get_object_or_404, redirect, render
+from django.utils import timezone
+
+from modulos.Authorization.decorators import permissions_required
+from modulos.Authorization.permissions import (POST_APPROVE_PERMISSION,
+                                               POST_CREATE_PERMISSION,
+                                               POST_DELETE_PERMISSION,
+                                               POST_EDIT_PERMISSION,
+                                               POST_EDIT_PERMISSION,
+                                               POST_PUBLISH_PERMISSION, POST_REJECT_PERMISSION, POST_REVIEW_PERMISSION)
+from modulos.Authorization.roles import ADMIN
 from modulos.Categories.models import Category
 from modulos.Posts.buscador import buscador
 from modulos.Posts.forms import NewPostForm, SearchPostForm
@@ -32,7 +36,7 @@ def home_view(req):
         HttpResponse: La respuesta HTTP con el contenido renderizado de la plantilla 'pages/home.html'.
     """
 
-    ctx = new_ctx(req, {"posts": Post.objects.all()[:10]})
+    ctx = new_ctx(req, {"posts": Post.objects.filter(status=Post.PUBLISHED)[:10]})
 
     return render(req, "pages/home.html", context=ctx)
 
@@ -52,6 +56,15 @@ def view_post(request, id):
     """
 
     post = get_object_or_404(Post, id=id)
+
+    # Permitir ver la publicación solo si está publicada o si el usuario es el autor o tiene permisos
+    if (
+        post.status != Post.PUBLISHED
+        and post.author != request.user
+        and not request.user.has_perm(POST_REVIEW_PERMISSION)
+    ):
+        return HttpResponseBadRequest("No tienes permiso para ver esta publicación.")
+
     # parse tags
     tags = post.tags.split(",") if post.tags else []
     tags = [tag.strip() for tag in tags]  # Remove leading/trailing whitespace
@@ -71,19 +84,22 @@ def view_post(request, id):
 @permissions_required([POST_CREATE_PERMISSION])
 def create_post(request):
     if request.method == "POST":
-        post = NewPostForm(request.POST, request.FILES)
-        if not post.is_valid():
-            print(post.is_valid())
-            print(post.errors)
+        form = NewPostForm(request.POST, request.FILES)
+        if not form.is_valid():
             return HttpResponseBadRequest("Datos proporcionados invalidos")
 
-        p = post.save(commit=False)
+        p = form.save(commit=False)
+        if "save_draft" in request.POST:
+            p.status = Post.DRAFT
+        elif "submit_review" in request.POST:
+            p.status = Post.PENDING_REVIEW  # Estado "Pendiente de Revisión"
+
         p.author = request.user
         p.save()
+
         return redirect("/posts/" + str(p.id))
 
     ctx = new_ctx(request, {"form": NewPostForm})
-
     return render(
         request,
         "pages/new_post.html",
@@ -105,21 +121,13 @@ def manage_post(request):
     else:
         posts = Post.objects.filter(author=request.user)
 
-    permisos = request.user.get_all_permissions()
-
-    # Definición de permisos en variables booleanas
-    perm_create = "UserProfile." + POST_CREATE_PERMISSION in permisos
-    perm_edit = "UserProfile." + POST_EDIT_PERMISSION in permisos
-    perm_delete = "UserProfile." + POST_DELETE_PERMISSION in permisos
-
-    # Definición de contexto basado en permisos
     ctx = new_ctx(
         request,
         {
             "posts": posts,
-            "perm_create": perm_create,
-            "perm_edit": perm_edit,
-            "perm_delete": perm_delete,
+            "perm_create": request.user.has_perm(POST_CREATE_PERMISSION),
+            "perm_edit": request.user.has_perm(POST_EDIT_PERMISSION),
+            "perm_delete": request.user.has_perm(POST_DELETE_PERMISSION),
         },
     )
 
@@ -128,9 +136,10 @@ def manage_post(request):
 
 # Vista para eliminar un post
 @login_required
-@permissions_required([POST_DELETE_PERMISSION])
 def delete_post(request, id):
-    post = get_object_or_404(Post, pk=id)
+    post: Post = get_object_or_404(Post, pk=id)
+    if not request.user.has_perm(POST_DELETE_PERMISSION) or post.author != request.user:
+        return HttpResponseForbidden("No tienes permiso para eliminar este post.")
 
     if request.method == "POST":
         post.delete()
@@ -143,13 +152,21 @@ def delete_post(request, id):
 
 # Vista para editar un post
 @login_required
-@permissions_required([POST_EDIT_PERMISSION, POST_CREATE_PERMISSION])
+@permissions_required([POST_EDIT_PERMISSION])
 def edit_post(request, id):
     post = get_object_or_404(Post, pk=id)
     if request.method == "POST":
         form = NewPostForm(request.POST, request.FILES, instance=post)
         if form.is_valid():
-            form.save()
+            p = form.save(commit=False)
+
+            if "save_draft" in request.POST:
+                p.status = Post.DRAFT
+            elif "submit_review" in request.POST:
+                p.status = Post.PENDING_REVIEW  # Estado "Pendiente de Revisión"
+
+            p.save()
+
             return redirect("post_list")
 
     form = NewPostForm(instance=post)
@@ -169,3 +186,77 @@ def search_post(request):
     else:
         # O redirige a donde sea apropiado si no hay búsqueda
         return redirect("home")
+
+
+# --------------------
+# Flujo de publicacion
+# --------------------
+
+
+@login_required
+@permissions_required([POST_CREATE_PERMISSION])
+def send_to_review(request, id):
+    post = get_object_or_404(Post, id=id)
+    post.status = Post.PENDING_REVIEW
+    post.save()
+    return redirect("kanban_board")
+
+
+@login_required
+@permissions_required([POST_APPROVE_PERMISSION])
+def aprove_post(request, id):
+    post = get_object_or_404(Post, id=id)
+    post.status = Post.PENDING_PUBLICATION
+    post.save()
+    return redirect("kanban_board")
+
+
+@login_required
+@permissions_required([POST_PUBLISH_PERMISSION])
+def publish_post(request, id):
+    post = get_object_or_404(Post, id=id)
+    post.status = Post.PUBLISHED
+    post.publication_date = timezone.now()
+    post.save()
+    return redirect("kanban_board")
+
+
+@login_required
+@permissions_required([POST_REJECT_PERMISSION])
+def reject_post(request, id):
+    post = get_object_or_404(Post, id=id)
+    post.status = Post.DRAFT
+    post.save()
+    return redirect("kanban_board")
+
+
+@login_required
+def kanban_board(request):
+    # Filtrar los posts según el estado
+    drafts = Post.objects.filter(status=Post.DRAFT, author=request.user)
+    pending_review = Post.objects.filter(status=Post.PENDING_REVIEW)
+    pending_publication = Post.objects.filter(status=Post.PENDING_PUBLICATION)
+    recently_published = Post.objects.filter(
+        status=Post.PUBLISHED,
+        publication_date__gte=timezone.now() - timedelta(days=5),
+    )
+
+    # Pasar las publicaciones a la plantilla para organizarlas en el tablero
+    return render(
+        request,
+        "pages/kanban_board.html",
+        new_ctx(
+            request,
+            {
+                "drafts": drafts,
+                "pending_review": pending_review,
+                "pending_publication": pending_publication,
+                "published": recently_published,
+                # permisos para mostrar los botones
+                "can_create": request.user.has_perm(POST_CREATE_PERMISSION),
+                "can_publish": request.user.has_perm(POST_PUBLISH_PERMISSION),
+                "can_approve": request.user.has_perm(POST_APPROVE_PERMISSION),
+                "can_reject": request.user.has_perm(POST_REJECT_PERMISSION),
+            },
+        ),
+    )
